+++ changelog.md
@@ -1,2 +1,2 @@
 ### Ore Growth 1.0.11
-- Initial release of Ore Growth for NeoForge
+- Initial release of Ore Growth for Minecraft 1.20.5 & 1.20.6